refactor(workflows): move release-docs helpers into .atomic/workflows/lib
Relocate the release-docs helper module from .atomic/workflow-utils/
to .atomic/workflows/lib/ and update the workflow and unit test
imports to match.

Assistant-model: Claude Fable 5
